+++ packages/drippie-mon/CHANGELOG.md
@@ -1,6 +1,27 @@
 # @eth-optimism/drippie-mon
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [9ba869a7]
+- Updated dependencies [050859fd]
+  - @eth-optimism/common-ts@0.3.1
+
+## 0.3.0
+
+### Minor Changes
+
+- 84a8934c: BaseServiceV2 exposes service name and version as standard synthetic metric
+
+### Patch Changes
+
+- Updated dependencies [d9e39931]
+- Updated dependencies [84a8934c]
+  - @eth-optimism/common-ts@0.3.0
+
 ## 0.2.0
+
 ### Minor Changes
 
 - 982cb980: Release drippie-mon